match: CGruntzCommand::SetMaskFromList count is u8, not i32 (EnqueueGroupCells -> 100 EXACT)
The body's own static_cast<u8>(count) / (count & 0xff) masks were the tell:
the parameter is a byte all the way up through CGruntzCmdMgr::EnqueueMulti,
so the u8 callers stop emitting a widening 'and ecx,0xff' at the call.
EnqueueGroupCells 96.39 -> 100, ReportRecordsA 93.41 -> 94.59,
ReportRecordsB 94.88 -> 96.43; EnqueueMulti/SetMaskFromList stay 100.

+++ include/Gruntz/GruntzCmdMgr.h
@@ -45,7 +45,7 @@ public:
     void EnqueueMulti(
         i32 enqueueFlag,
         char targetIndex,
-        i32 count,
+        u8 count,
         u8* gruntList,
         char cmdKind,
         i16 posX,
